requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
pdfminer.six
scikit-learn
nltk
tqdm
